Add permissions. Closes #142.

+++ src/Http/Controllers/ReportController.php
@@ -3,6 +3,7 @@
 namespace Statamic\SeoPro\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Statamic\Facades\User;
 use Statamic\Http\Controllers\CP\CpController;
 use Statamic\SeoPro\Reporting\Report;
 
@@ -10,6 +11,8 @@ class ReportController extends CpController
 {
     public function index(Request $request)
     {
+        abort_unless(User::current()->can('view seo reports'), 403);
+
         if (! $request->ajax()) {
             return view('seo-pro::reports');
         }
@@ -19,16 +22,22 @@ class ReportController extends CpController
 
     public function store()
     {
+        abort_unless(User::current()->can('view seo reports'), 403);
+
         return Report::queue();
     }
 
     public function show(Request $request, $id)
     {
+        abort_unless(User::current()->can('view seo reports'), 403);
+
         return Report::find($id)->withPages();
     }
 
     public function destroy($id)
     {
+        abort_unless(User::current()->can('delete seo reports'), 403);
+
         return Report::find($id)->delete();
     }
 }
